+++ tests/unit/Docs/PersonRequestTest.php
@@ -59,6 +59,20 @@ class PersonRequestTest extends TestCaseHttpMocked
     /**
      * @doesNotPerformAssertions
      */
+    public function testCreatePersonWithEqualName()
+    {
+        $newPerson = new Person();
+        $newPerson->setFirstName("John")
+            ->setLastName("Doe")
+            ->setBirthday("1970-01-01");
+        //add further attributes
+
+        PersonRequest::create($newPerson, force: true);
+    }
+
+    /**
+     * @doesNotPerformAssertions
+     */
     public function testUpdatePerson()
     {
         $person = PersonRequest::findOrFail(21);
@@ -116,5 +130,32 @@ class PersonRequestTest extends TestCaseHttpMocked
         $this->assertEquals("1997-03-01", $lastBirthdayPerson->getAnniversaryInitialDate());
         $this->assertEquals("2022-03-01", $lastBirthdayPerson->getAnniversary());
         $this->assertEquals("25", $lastBirthdayPerson->getAge());
+    }
+
+    public function testRequestTags()
+    {
+        $person = (new Person())->setId("21");
+
+        $tags = $person->requestTags()?->get();
+
+        if ($tags == null) {
+            $tags = [];
+        }
+
+        $musicDirectorTag = null;
+        foreach ($tags as $tag) {
+            if ($tag->getName() == "Music Director") {
+                $musicDirectorTag = $tag;
+            }
+        }
+        // Tag-Data
+        $this->assertEquals(5, $musicDirectorTag?->getId());
+        $this->assertEquals("Music Director", $musicDirectorTag?->getName());
+        $this->assertEquals(9, $musicDirectorTag?->getCount());
+
+        // Meta-Data
+        $this->assertEquals("2021-05-19T06:21:02Z", $musicDirectorTag?->getModifiedAt());
+        $this->assertEquals(21, $musicDirectorTag?->getModifiedBy());
+        $this->assertEquals("Matthew", $musicDirectorTag?->requestModifier()?->getFirstName());
     }
 }